Added some #define macros for some predefined values that are determined by the dev.
Made the roi struct a Class instead, at this point it makes more sense.

Added smoothing to the occupied status via the setOccupied method. This method does not set the occupied instance variable to isOccupied. It collects N number of isOccupied records then determines if more than half of them have a value of true, then it will set the occupied instance variable to true or false. The purpose of this is to "smooth" out the occupied variable since noise can briefly make an ROI that is on the edge of being occupied/not occupied change statuses for a frame or two.